pyproject.toml
[tool.poetry]
name = "flirextractor"
version = "0.1.0"
description = "A library from extracting temperature values from FLIR IRT Images."
authors = ["Alois Klink <alois.klink@gmail.com>"]
license = "GPL-3.0+"
readme = "README.md"
repository = "https://github.com/aloisklink/flirextractor"
homepage = "https://github.com/aloisklink/flirextractor"
classifiers = [
    "INTENDED AUDIENCE :: SCIENCE/RESEARCH",
    "TOPIC :: MULTIMEDIA :: GRAPHICS :: CAPTURE :: DIGITAL CAMERA",
]

[tool.poetry.dependencies]
python = "^3.6"
pyexiftool = "^0.1.1"
pillow = "^6.2"
numpy = "^1.17"

[tool.poetry.dev-dependencies]
black = {git = "https://github.com/psf/black.git", tag = "19.3b0"}
mypy = "^0.730.0"
flake8 = "^3.7"
flake8-mypy = "^17.8"
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
